Moved NewEdmObject
