+++ src/views/Main/Schemas/Components/NewEdmObjectInput.js
@@ -1,9 +1,9 @@
 import React, { PropTypes } from 'react';
 import { EntityDataModelApi } from 'loom-data';
-import { NameNamespaceAutosuggest } from '../namespaceautosuggest/NameNamespaceAutosuggest';
-import StringConsts from '../../utils/Consts/StringConsts';
-import EdmConsts from '../../utils/Consts/EdmConsts';
-import styles from '../../views/Main/Schemas/styles.module.css';
+import { NameNamespaceAutosuggest } from '../../../../components/namespaceautosuggest/NameNamespaceAutosuggest';
+import StringConsts from '../../../../utils/Consts/StringConsts';
+import EdmConsts from '../../../../utils/Consts/EdmConsts';
+import styles from '../styles.module.css';
 
 const NAME_FIELD = 'name';
 const NAMESPACE_FIELD = 'namespace';
