+++ src/views/Main/Schemas/Components/SchemaList.js
@@ -3,7 +3,7 @@ import { EntityDataModelApi } from 'loom-data';
 import { Promise } from 'bluebird';
 import Utils from '../../../../utils/Utils';
 import { Schema } from './Schema';
-import { NewEdmObjectInput } from '../../../../components/edminput/NewEdmObjectInput';
+import { NewEdmObjectInput } from './NewEdmObjectInput';
 import EdmConsts from '../../../../utils/Consts/EdmConsts';
 import styles from '../styles.module.css';
